% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfWheel.R
\name{pdfWheel}
\alias{pdfWheel}
\title{pdfWheel; an efficient PDF embedder for Shiny and RMarkdown}
\usage{
pdfWheel(
  name,
  path,
  buttons = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{path}{a character, path to a directory containing the pages of the PDF you wish to embed. File names should be "page_x.jpg" where "x" is the 1-indexed page number.}

\item{buttons}{a boolean, do you want next/previous page buttons beneath the embedding?}
}
\description{
<Add Description>
}
